fix button with links for navigation
Buttons that were used for navigation that were not part of a form
used anchor tags to provde navigation. This was clunky and made the
button not work right when the user didn't click the words. Now, button
will be the link and navigate to the correct page on click.